% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multimark_functions.R
\name{getfreq}
\alias{getfreq}
\title{Generate encounter history frequencies}
\usage{
getfreq(Enc.Mat, data.type)
}
\arguments{
\item{Enc.Mat}{A matrix of observed encounter histories with rows corresponding to individuals and columns corresponding to sampling occasions.}

\item{data.type}{Specifies the encounter history data type. All data types include non-detections (type 0 encounter), type 1 encounter (e.g., left-side), and type 2 encounters (e.g., right-side). When both type 1 and type 2 encounters occur for the same individual within a sampling occasion, these can either be "non-simultaneous" (type 3 encounter) or "simultaneous" (type 4 encounter). Three data types are currently permitted:

 \code{data.type="never"} indicates both type 1 and type 2 encounters are never observed for the same individual within a sampling occasion, and observed encounter histories therefore include only type 1 or type 2 encounters (e.g., only left- and right-sided photographs were collected). Observed encounter histories can consist of non-detections (0), type 1 encounters (1), and type 2 encounters (2). See \code{\link{bobcat}}. Latent encounter histories consist of non-detections (0), type 1 encounters (1), type 2 encounters (2), and type 3 encounters (3).

 \code{data.type="sometimes"} indicates both type 1 and type 2 encounters are sometimes observed (e.g., both-sided photographs are sometimes obtained, but not necessarily for all individuals). Observed encounter histories can consist of non-detections (0), type 1 encounters (1), type 2 encounters (2), type 3 encounters (3), and type 4 encounters (4). Type 3 encounters can only be observed when an individual has at least one type 4 encounter. Latent encounter histories consist of non-detections (0), type 1 encounters (1), type 2 encounters (2), type 3 encounters (3), and type 4 encounters (4).

 \code{data.type="always"} indicates both type 1 and type 2 encounters are always observed, but some encounter histories may still include only type 1 or type 2 encounters. Observed encounter histories can consist of non-detections (0), type 1 encounters (1), type 2 encounters (2), and type 4 encounters (4). Latent encounter histories consist of non-detections (0), type 1 encounters (1), type 2 encounters (2), and type 4 encounters (4).}
}
\value{
A vector of encounter history frequencies of length \eqn{4^T} for the "never" and "always" data types or of length \eqn{5^T} for the "sometimes" data type.
}
\description{
This function generates an ordered vector of encounter history frequencies
}
\details{
This function calculates frequencies for the latent encounter history frequencies by assuming a 1-to-1 mapping of the observed histories (\code{Enc.Mat}) to latent histories.

For the \code{"never"} and \code{"always"} data types, the index of latent history \eqn{i = 1 + \sum_{t=1}^T \omega_t*4^{T-t}}{i = 1 + \sum_{t=1}^T \omega_t*4^{T-t}}, with \eqn{\omega_t \in {0, 1, 2, 3}} to represent a non-encounter, a type 1 encounter, a type 2 encounter, and both type 1 and type 2 encounter (i.e., type 3 encounter for \code{"never"} and type 4 encounter for \code{"always"}), respectively, for sampling occasion \eqn{t=1,\ldots,T}.

For the "sometimes" data type, the index of latent history \eqn{i = 1 + \sum_{t=1}^T \omega_t*5^{T-t}}{i = 1 + \sum_{t=1}^T \omega_t*5^{T-t}}, with \eqn{\omega_t \in {0, 1, 2, 3,  4}} to represent a non-encounter, a type 1 encounter, a type 2 encounter, type 3 encounter, and type 4 encounter, respectively, for sampling occasion \eqn{t=1,\ldots,T}.
}
\examples{
#simulate closed capture data with 2 sampling occasions for data.type="sometimes"
data_type <- "sometimes"
data <- simdataClosed(noccas=2,data.type=data_type)

#generate initial latent history freqencies from observed encounter histories
getfreq(data$Enc.Mat,data.type=data_type)

#calculate freqencies for true encounter histories
getfreq(data$trueEnc.Mat,data.type=data_type)
}
\author{
Brett T. McClintock
}
\references{
Bonner, S. J., and Holmberg J. 2013. Mark-recapture with multiple, non-invasive marks. \emph{Biometrics} 69: 766-775.

McClintock, B. T., Conn, P. B., Alonso, R. S., and Crooks, K. R. 2013. Integrated modeling of bilateral photo-identification data in mark-recapture analyses. \emph{Ecology} 94: 1464-1471.
}

